Added support to pom.xml to automatically pull in static web content. Currently, anything in ../../javascript with either an "html", "css", or "js" file extension will be included, including paths. This may be changed to be an explicit list of files instead of pulling everything in the directory.
This change will only work if you actually use a jar or war inside of a servlet container that is properly configured to serve static files out of the root of the jar. This behavior comes automatically with jetty, but you must use the jetty:run-war goal and not jetty:run, because jetty:run examines files on the local file system in src/main/webapp.

Do NOT move static content files into src/main/webapp directory. These files are shared with other language implementations and do not belong in the java code base.



git-svn-id: http://svn.apache.org/repos/asf/incubator/shindig/trunk@611391 13f79535-47bb-0310-9956-ffa450edef68