Remove references to signing and add URL to mcrypt library
According to http://www.apache.org/dev/crypto.html#faq and
irc conversations, signatures are not considered crypto to
the effect of notifications. Remove references to them, and
thus OAuth from the NOTICE. Keep AES, the reference to JCE
for Java and the reference to mcrypt as the software linked to
in the PHP case.


git-svn-id: http://svn.apache.org/repos/asf/incubator/shindig/trunk@655316 13f79535-47bb-0310-9956-ffa450edef68