SHINDIG-146: Rename remaining "synd" -> "cont"
There was a remaining reference to syndicator
in the secureToken used if no one is present, by
gadget.js. I assume it is a dummy word, as there is
no code matching it anywhere in the rest of the code
and a question in the list to the effect never got
any answer.


git-svn-id: http://svn.apache.org/repos/asf/incubator/shindig/trunk@655318 13f79535-47bb-0310-9956-ffa450edef68