Simplify BasicHttpCacheTest to use HttpUtil to parse Expiry date
After the patch for SHINDIG-257 was applied, just this one
use of SimpleDateFormat was remaining, and it actually caused a bug
for non English locale users before it was patched. This patch
makes it simpler and it will also make easier doing the right thing
for people looking into the tests to learn how to use the code.


git-svn-id: http://svn.apache.org/repos/asf/incubator/shindig/trunk@656561 13f79535-47bb-0310-9956-ffa450edef68