Same-domain gadgets.rpc call support.
Detects if the receiving context of a given gadgets.rpc call
is on the same domain as the caller. In that case, makes a direct call to the
receiver rather than relying on other cross-domain tricks. This is useful
for trusted gadgets rendered on the same domain as their container, irrespective
of browser, since this "transport" is simply a function call that works on any
browser.

This patch closes SHINDIG-456.



git-svn-id: http://svn.apache.org/repos/asf/incubator/shindig/trunk@677755 13f79535-47bb-0310-9956-ffa450edef68