Unbound the DataServiceServlet BeanConverters from their implementations and rebound them to the Interface. Modified the injected setter to use named beans and updated the relevant modules to inject the same implementations as before.
This change will allow alternative implementations of a BeanConverter to be injected by Guice, and avoid binding
to implementations. There is no functional change created by this patch. About 10 lines in total.

I am assuming commits like this, which dont generate functional change  are small enough not to require
JIRA discussion, but please tell me if this is not the case ?

Very happy to revert/modify if required, just let me know.


git-svn-id: http://svn.apache.org/repos/asf/incubator/shindig/trunk@678290 13f79535-47bb-0310-9956-ffa450edef68