Second try.
Resolution to SHINDIG-501.

Framework for generating a parse tree of HTML and CSS content.

These interfaces are defined in order to cleanly separate parsing logic
from classes that manipulate a given parse tree (HTML or CSS). The parse
tree objects are intended to be only as complex as is needed for the vast
majority of content rewriting manipulation. They provide structure but
no more semantics (validation, CSS resolution, etc.) than that.

This code is new, and does not affect existing functionality.



git-svn-id: http://svn.apache.org/repos/asf/incubator/shindig/trunk@686103 13f79535-47bb-0310-9956-ffa450edef68