Fix for SHINDIG-555.
This patch eliminates the remaining ad hoc security token handling. All auth is now done via the auth filter.

Due to the inability to overwrite bindings in the current version of Guice, I have added a small, auth-only module for gadget rendering that is used when running a standalone gadget rendering server. For full or social only servers, the social binding is used. The gadget binding does not include oauth as a valid auth handler.

With this change complete, many other pieces of code can probably be removed. SecurityTokenDecoder is most likely an obsolete interface.



git-svn-id: http://svn.apache.org/repos/asf/incubator/shindig/trunk@691158 13f79535-47bb-0310-9956-ffa450edef68