Extracted interface for ContainerConfig to allow implementations to choose alternative configuration mechanisms if their deployment needs them.
We should probably create more generic primitive wrappers for the JSONObject / JSONArray, but that would be a very disruptive change.

Since Java 1.6 has native support for javascript, I'd also like to add a ContainerConfig variant that allows use of full javascript, making inheritance easier as well as allowing for imports and expressions. Since this would be a sharp departure from compatibility with the PHP code, that would be a strictly separate binding. Maintaining compatibility with the client code is of the utmost importance.



git-svn-id: http://svn.apache.org/repos/asf/incubator/shindig/trunk@706822 13f79535-47bb-0310-9956-ffa450edef68