This is an initial implementation of the "Limited Invalidation" spec. Containers are expected to implement their own InvalidationService based on their caching behavior
This implementation differs from the spec in a couple of ways that I believe are
worth proposing as clarifications.
- The service endpoint is "cache/invalidate" not "invalidate" for REST. For RPC
the method is "cache.invalidate"
- Only 2-legged consumer OAuth calls can invalidate URLs
- Security token for rendered gadgets can only be used to invalidate the viewers data
-  Viewer data is invalidated by default and a simple GET /social/rest/cache/invalidate will trigger it

The default request pipeline is altered to allow invalidated content to be
returned if fetching a replacement causes an error as is suggested in the spec.

git-svn-id: http://svn.apache.org/repos/asf/incubator/shindig/trunk@749483 13f79535-47bb-0310-9956-ffa450edef68