This patch changes the gadget rendering to a completely DOMDocument model (using the GadgetRewriter class), and now adds the javascript and css blobs via the DOM.
Because of this the same dom actions can be re-used by the HrefRenderer in order to have a valid and consistent
style and gadgets API available there too.

On href views (proxied content) it automatically append a setHeight() call to the init script, so the view is
always fully visible

The only thing still missing from the proxied content/data-pipelining implementation now is the EL expression
parsing, which will be dealt with by the to-be-implemented template parser; Otherwise it's now feature complete.



git-svn-id: http://svn.apache.org/repos/asf/incubator/shindig/trunk@764191 13f79535-47bb-0310-9956-ffa450edef68