Fix AbstractTagHandler.getValueFromTag() to return null if the attribute isn't present.
The existing behavior led to lots of unnecessary expression evaluation, in particular unnecessary coercion, though not returning null is kinda handy.  Added a test for AbstractTagHandler, as well as tests for IfTagHandler and RepeatTagHandler that verify behavior when their relevant attributes are altogether missing.  And make RepeatTagHandler not depend on DefaultTemplateProcessor ignoring null Iterables for repeating.


git-svn-id: http://svn.apache.org/repos/asf/incubator/shindig/trunk@790719 13f79535-47bb-0310-9956-ffa450edef68