Tiny refactoring of HtmlRenderer code.
Full disclosure: I am committing this with the intent to, in our build of Shindig, subclass HtmlRenderer to support a composite "test"-type view. The general idea is, for a render with query param &testmode=1 (or similar), view content (proxy-mode not supported initially) will come from the "host" view but with content from "test:<host>" view, if any, appended to it. The test view may in turn do whatever it wants. We'll render gadgets that support such views in our test harness, with success criteria being:
A) no uncaught exceptions were thrown
B) a special gadgets.rpc ALL_OK callback handler is fired with an affirmative value of success.

Because this will be purely experimental and the details will undoubtedly change, I'm keeping the hook as minimal as possible in core shindig code.



git-svn-id: http://svn.apache.org/repos/asf/incubator/shindig/trunk@886909 13f79535-47bb-0310-9956-ffa450edef68